requirements.txt
streamlit
python-pptx
openai
Pillow
